--- src/common/history.h
@@ -1,14 +1,0 @@
-
-#pragma once
-
-#include <string>
-
-
-namespace History {
-
-extern std::string filename;
-
-void setup();
-void append(char *line);
-
-} // namespace History
