+++ src/history.h
@@ -1,0 +1,9 @@
+
+#pragma once
+
+#include <string>
+
+namespace History {
+	void init();
+	void append(char *line);
+}
