+++ src/main.cpp
@@ -1,11 +1,11 @@
 
 #include "emulator.h"
+#include "history.h"
 #include "common/logger.h"
-#include "common/history.h"
 
 int main(int argc, const char *argv[]) {
 	Logger::init(Logger::DEBUG);
-	History::setup();
+	History::init();
 
 	Emulator *emulator = new Emulator();
 	emulator->run();
